requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
chromadb==0.4.22
openai==1.10.0
python-dotenv==1.0.0
streamlit==1.30.0
pypdf==3.17.4